fix(niri): decode a sequence length at its own width, and distill the branch
`Value::Int` holds the sign-extended bit pattern, so a negative `i32` length
read back as a huge `u64` and reached the element-count check instead of the
refusal meant for it. The trace said "used <huge> of N elements" where it
meant "negative len". Behaviour is unchanged: both paths refuse.

No test: a negative length cannot arise from a well-formed program, so the
branch is a defensive guard and a fixture for it would assert what it sets up.

Distilled: the derived materializing-global set now says why it is rebuilt on
every call — it reads body contents, which the count-keyed fold cache does not
track, so moving it into that cache would leave a stale set behind. The WEP
lists `region_seed` beside the two instruments it belongs with, and three
sentences that a reader had to decode are plain.

+++ wado-compiler/src/niri/rewrite.rs
@@ -217,7 +217,9 @@ impl Interpreter<'_> {
             );
             return false;
         };
-        let Ok(used) = usize::try_from(used) else {
+        // `Value::Int` holds the sign-extended bit pattern, so a negative length
+        // reads as a huge `u64` until it is decoded at its own width.
+        let Ok(used) = usize::try_from(used as i32) else {
             crate::compiler_trace!("region_seed", "materialize {e:?}: negative len {used}");
             return false;
         };
